Updated MockInterface inline with RHInterface.
Reduced code-duplication and increased testability.
